// src/@shared/models/internal-user-profile.ts
export interface InternalUserProfile {
    ID: number;
    UserId: string;
    FirstName: string;
    LastName: string;
    PrimaryEmail: string;
    PrimaryPhoneNo: string;
    PrimaryPhoneExt: string;
    Password: string;
    SecondaryEmail: string;
    SecondaryPhoneNo: string;
    SecondaryPhoneExt: string;
    SecondaryPhoneCountryExt:string;
    PrimaryPhoneCountryExt:string;
    Address1: string;
    Address2: string;
    City: string;
    State: string;
    Zip: any;
    CompanyName: string;
    SkillSet: string;
    ProfilePic: string;
    IsActive: boolean;
    UpdatedDate: Date;
    CreatedDate: Date;
    LastLoggedIn: Date;
    EmployerType: number;
    Website: string;
    IsRecruiter: boolean;
    IsScheduler: boolean;
    IsInterviewer: boolean;
    IsRequestPending:boolean;
    Description:string;
    KeyWords:string[];
    CompanyId:number;
    Logo:string;
    TimeZoneId:number;
    PrimaryCountryCode:string;
    SecondaryCountryCode:string;
    TimeZoneName:string;
}